revamped training folder to be "standalone"
instead of being built with rest of application with CMake, Cython will be used to build C++ code that interfaces with the python training code.

Also started implementing gtests
