+++ src/chess/CMakeLists.txt
@@ -5,5 +5,3 @@ file(GLOB HEADER_FILES CONFIGURE_DEPENDS "*.h")
 file(GLOB SRC_FILES CONFIGURE_DEPENDS "*.cpp")
 
 add_library(dupchess_chess STATIC ${SRC_FILES} ${HEADER_FILES})
-
-target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
