+++ tests/tests.cpp
@@ -1,0 +1,11 @@
+#include <gtest/gtest.h>
+
+
+
+// Demonstrate some basic assertions.
+TEST(HelloTest, BasicAssertions) {
+	// Expect two strings not to be equal.
+	EXPECT_STRNE("hello", "world");
+	// Expect equality.
+	EXPECT_EQ(7 * 6, 42);
+}
